feat(metrics): extract PR #1175 V3 codec into standalone datadog-metrics-v3 crate
After rebasing stephen/v3_vector on tobz/datadog-metrics-v3-payload-support
(PR #1175), extract the codec internals into lib/datadog-metrics-v3 so
Vector can reuse them without depending on saluki-core or saluki-common.

Extracted into standalone crate (zero external deps):
- Interner: Borrow-trait based get_or_insert — no clone on cache hits
- Types: V3MetricType, V3ValueType, PointKind, value_type_for_values
  (the precision-safe compaction algorithm from PR #1175)
- Writer: unit support (FLAG_HAS_UNIT / dict_unit_bytes / unit_refs),
  FLAG_NO_INDEX, set_unit(), set_no_index()
- Serializer: fields 25 (dictUnitStr) and 26 (unitRefs)

Removed from saluki-components (now re-exported from standalone crate):
- v3/interner.rs → deleted; use datadog_metrics_v3::Interner
- v3/types.rs    → deleted; use datadog_metrics_v3::{V3MetricType, V3ValueType, ...}
- v3/writer.rs   → kept (saluki-specific output types + protobuf serialization)
  but now imports Interner, V3MetricType, V3ValueType from the standalone crate

The saluki-components v3/writer.rs retains its own value_type_for_values
(identical algorithm) and protobuf::CodedOutputStream serialization since
those are appropriate for saluki. The standalone crate uses a zero-dep
hand-rolled serializer for Vector compatibility.
