+++ lib/datadog-metrics-v3/src/lib.rs
@@ -1,8 +1,9 @@
-mod interner;
+pub mod interner;
 mod serializer;
 mod types;
 mod writer;
 
+pub use interner::Interner;
 pub use serializer::serialize_v3_payload;
-pub use types::{V3MetricType, V3ValueType};
+pub use types::{value_type_for_values, FLAG_HAS_UNIT, FLAG_NO_INDEX, V3MetricType, V3ValueType};
 pub use writer::{V3EncodedData, V3MetricBuilder, V3Writer};
