+++ lib/saluki-components/src/encoders/datadog/metrics/v3/mod.rs
@@ -16,14 +16,12 @@
 //!   but we do this in a single shot as part of this initial implementation.
 
 mod constants;
-mod interner;
 mod payload;
 mod telemetry;
-mod types;
 mod writer;
 
 pub(super) use payload::{V3EncodedRequest, V3PayloadLimits, V3PayloadRequest};
 pub(crate) use telemetry::{V3PayloadSplitReason, V3SerializerTelemetry};
-pub use types::V3MetricType;
+pub use datadog_metrics_v3::V3MetricType;
 pub(crate) use writer::V3EncoderStats;
 pub use writer::{V3EncodedMetrics, V3Writer};
